allow PaymentsDetailsResponse to deserialize responses that include an 'action'
