+++ src/test/resources/mocks/checkout/paymentsdetails-challenge-shopper.json
@@ -1,0 +1,11 @@
+{
+  "resultCode": "ChallengeShopper",
+  "action": {
+    "paymentData": "Ab02b4c0!BQABAgA4e3wGkhVah4CJL19qdegdmm9E...",
+    "paymentMethodType": "scheme",
+    "authorisationToken": "BQABAQBNkShWvWlisqJoxqmswYKscu+WOrAczR0QrZgB6qesj012LDPcjvJqbhurlRgPVwt6uis2+i7lwtQCWtXzUb+hf6SmB28gOxN1GIkfsvhKfPuZPbMeFtU4N0CiJ/6QRqoSOkJn2Lbutqbx1MaQmyCR6g0J8SZZPpRdg0vBVW/3QcOSziEsdaKVrb2dz+q2VUD6eO/aHBLU/vKD0RUxEm+4MY+HkmNrZ+BmXcdSeV+829YlU3iX4NKsqvB6YmxohQoCUvNrnpX9VhfJtyzxdtY5kY//GqlES05q+IY0QBTmZRw8i6nM7Xc+C5ID03xrMXw4AmCIidfhX94GcD0A93izECEDNFomaeCh+flQpARKA98AAHMEjM1yxMYuByKHdwAa+CHiSPKhfTZnB9CmL5uZAmEZy4vrGpXXlpvFEoKndHUjKAbwZiVqPRAL3PG5h6bQeffRZJM683kdHwWUv+tzMKWnyrS8NeI1gjmpwaanHSysERSvB10vo1tXBtZSahHxCuHgbNmQo/4cdZKWPKnFnc8BZ4UH4TWvwBhMR4KtUkgN9TTBhXbXbjCdMXkDCb7ie0OibBt/a190HFxOiLvC/kmTiQAxFu4BJac7n6gt+1XOX+eU55h8UiV3HJo9HkGE6gBgt/Z9FKuQ4WYBVteHqeulWim3w3Pwt+zYlZzx33TXtK1YLGvZn5i6TZMlOOmklZGOONGkIS0j9Z3dB+ub2f14h7/bVv2vLgbLYs7M/Ndysf52q6nqmCEkdwHCRRcdXZwrZIvMobdO69lOYzMtmMe/dcyMIjKrpeoia62t1jIQdfIy8Ov28JCJyjqacthn4tGcBZze5dI25I67VvCt34zw2Q+gZgJsug4GD4WjCCqJH479CzAMGxdnLejDz7Jjv1AbFxe4qKkkQ26A178lwmt+TE8tP+Hfl7TJ4o5wDzIv1Rrau4hq51IZdOlkRah079IZwwQv3A32VKRcv68p/ANa5kcKn3WDPA9WZ1HBfkgMotjZvZTFj6Kr1QMtOfpGA3Wm3nNJm8NMj5r80czldtYfIEp19Bcn4z2idh6VCiu6DTyOduRssD5Dt4CdoHhlpEuBcU27LYgGaZOSfQ8bPR/IUQTlEbByJ66Xp/++a0vvs6fo1eURsw5Oqpn7eOUHJOPMBHRZ9aQDhsDnjI8HbVJ/ikbTGBhTRc8qrKrksxEh2lWi9QiTBXNsNjlPySyPxcBR5c/WuQoMQZiY8ePwWVKoOpNXH17epmT7LbgC1aCPi5efNs7qsrvwsCmV1w6+sIpq4PPRL2DHFvo4KZK5aZP3AK2e7CSHkUcHXLguwmoHW3e7QZ0JvF/HWRTla4Ayj76fuvwUQGD9xzjSJgHYXJ/veOsHXet3lAhVXOU2XA18cxyLjcug1t3nGc0AvUQrwJ3PpNStR6UwRUVAuM/O+G5hb+cR03u5q/eUYG3V9mkT3ql+/dY8TSXuyFtw51T5LxvIJUhuWrV4kR0D+Dq/D6Zs/m9h5O74MesR6efHuGKoV/JNzNqd1T8qBwD1kS+qtoXLBRvDSjrXdZ2xBhrv/BGQFH+Opcj0DkVReoHu7RzXbrg2d6UaNcHmjYjx3fHXqmFRxkzEIeEBKfLy7IjoTyXFcgDUt/qFd0GaiNMCN2aXkChtEoIGQ5Ik8tkJYIPSKfjk8bUlBeUvyMZ2Z6Pt8e/wic86RkJl6f70qIWbQoaevaYh3V0LNoUhlAqIWVKNbqkecHH7eAI7pdrjzPgEY7MvvFHV6QN5SGd8tZXJg84XMZNp7faprCsJdR43fntvfvMHXX1KhpEiVX60r8LaAxgsVJPgFYchFw5wbkSvF2XbGpNsbPhpU2x9WDcj34nS1a8pydGxd0658HYVqwYjFE/5LCtMNxgU5HZU4ekdbbL6hEIDgcql9zmJXlWdvSdKJZ+4TcKiqKE/bXBgHF4Ip7L6eo8V0UNPOvXOU66+CoKz3Xfc5Qip2vL18ff5heVPbl//njECE0Q3BX+m3AK7ugLGVjpsMCMjhgHr+8eEvXML1wch5TEw663+AHFRYHzkCX6mllF/llLkdwDfTBSmTa24ukRXaolSvi1hcoPLN8GpjlSfrLjY5fpmS1eqeGSZuZuYmhJZxpm1APzPcHJVseNhzU6sodfelzxVoJ03qHLYeVLoFLkuAfLEXGbRAMGfcUVSz6elm7dOO8Z5pQ9nXgYeBJd95o2GoPZ5Xpl8zqs51Fc0iLY55fehEDDtTSy0vqcKO3KOnb23hrxyV8VStLxEY5kR",
+    "subtype": "challenge",
+    "token": "eyJ0aHJlZURTTWVzc2FnZVZlcnNpb24iOiIyLjEuMCIsInRocmVlRFNNZXRob2ROb3RpZmljYXRpb25VUkwiOiJodHRwczpcL1wvY2hlY2tvdXRzaG9wcGVyLXRlc3QuYWR5ZW4uY29tXC9jaGVja291dHNob3BwZXJcL3RocmVlRFNNZXRob2ROb3RpZmljYXRpb24uc2h0bWw/b3JpZ2luS2V5PXB1Yi52Mi44MjE2MTUzNzkyOTgzMTQxLmFIUjBjRG92TDJ4dlkyRnNhRzl6ZERvNU1EQXcuN2dUdk42UjlmaVFRX3RxbzlRVnBZR0Y3U240V2w0MEpxdFlPN0NiS0pJayIsInRocmVlRFNNZXRob2RVcmwiOiJodHRwczpcL1wvcGFsLXRlc3QuYWR5ZW4uY29tXC90aHJlZWRzMnNpbXVsYXRvclwvYWNzXC9zdGFydE1ldGhvZC5zaHRtbCIsInRocmVlRFNTZXJ2ZXJUcmFuc0lEIjoiNDg5Njc1MGMtMDYyZS00ZjgyLTk0ZTUtMTI2ZWJiNWJkNWE0In0=",
+    "type": "threeDS2"
+  }
+}
